refactor: load the homepage in Next.js
Bootstrap and custom css styles are applied in the Next.js structure, so
that the routing and component refactoring can be done separately from
the style refactoring. Each page or component's styling must be
refactored in order to be loaded in Next.js app, and the homepage is the
first one refactored.

TODO: The image carousel in the homepage, which uses bootstrap, is
broken. It is advised to use a separate library instead of bootstrap for
this purpose.